blockchain node initialization sample code
This code is the main function for my Node.c file which initializes the BC node on the Linux server. More specifically, the code contains the main function as well as the function that works on the thread dealing with inbound messages about the BC system from other nodes- these messages include messages of node initialization, notices of other nodes finding the correct PoW, consensus messages, and more.